Moved the common frontend authentication code into its own class
