+++ src/CodingMonkey/src/admin/exercise-category/details.js
@@ -3,18 +3,15 @@ import {HttpClient, json} from 'aurelia-fetch-client';
 import {Router} from 'aurelia-router';
 import 'fetch';
 import toastr from 'toastr';
+import {Authentication} from './../../authentication/authentication.js';
 
 @inject(HttpClient, Router)
 export class details {
     constructor(http, router) {
         this.appRouter = router;
+
         // Check there is a logged in user or kick them to homepage
-        var currentUser = this.getCurrentUserInSessionStorage();
-
-        if (!currentUser.isLoggedIn) {
-            this.appRouter.navigate("/");
-        }
-
+        new Authentication(this.appRouter).verifyUserLoggedIn();
         var loc = window.location;
         
         this.heading = "Exercise Category Details";
@@ -86,12 +83,5 @@ export class details {
     
     goToExerciseCategoryList() {
         this.appRouter.navigate("admin/exercise/categories");
-    }
-
-    getCurrentUserInSessionStorage() {
-        let currentUserRaw = sessionStorage.getItem("currentUser");
-        let currentUser = JSON.parse(currentUserRaw);
-
-        return currentUser;
     }
 }
