+++ src/CodingMonkey/src/admin/test/create.js
@@ -3,17 +3,15 @@ import {HttpClient, json} from 'aurelia-fetch-client';
 import {Router} from 'aurelia-router';
 import 'fetch';
 import toastr from 'toastr';
+import {Authentication} from './../../authentication/authentication.js';
 
 @inject(HttpClient, Router)
 export class create {
     constructor(http, router) {
         this.appRouter = router;
-        // Check there is a logged in user or kick them to homepage
-        var currentUser = this.getCurrentUserInSessionStorage();
 
-        if (!currentUser.isLoggedIn) {
-            this.appRouter.navigate("/");
-        }
+        // Check there is a logged in user or kick them to homepage
+        new Authentication(this.appRouter).verifyUserLoggedIn();
 
         var loc = window.location;
         
@@ -165,12 +163,5 @@ export class create {
     
     goToTestList() {
         this.appRouter.navigate("admin/exercise/" + this.vm.exercise.id + "/tests");
-    }
-
-    getCurrentUserInSessionStorage() {
-        let currentUserRaw = sessionStorage.getItem("currentUser");
-        let currentUser = JSON.parse(currentUserRaw);
-
-        return currentUser;
     }
 }
