+++ src/CodingMonkey/src/main.js
@@ -3,7 +3,6 @@ import 'bootstrap';
 export function configure(aurelia) {
   aurelia.use
     .standardConfiguration()
-    .developmentLogging()
     .plugin('aurelia-dialog');
 
   //Uncomment the line below to enable animation.
